app/requirements.txt
flask==1.0.2
flask_httpauth
itsdangerous
pyYAML
pyjwt
datetime
config